+++ src/Components/Publishing/Nav/NewsNav.tsx
@@ -1,8 +1,8 @@
-import moment from "moment"
 import React from "react"
 import styled from "styled-components"
+import colors from "../../../Assets/Colors"
 import { pMedia } from "../../Helpers"
-import { Fonts } from "../Fonts"
+import { NewsDateHeader, NewsText } from "../News/NewsDateHeader"
 
 interface Props {
   date: string
@@ -10,30 +10,17 @@ interface Props {
 
 export const NewsNav: React.SFC<Props> = props => {
   const { date } = props
-  const today = new Date()
-  const hasYear = moment(date).format("YYYY") !== moment(today).format("YYYY")
-  const isToday =
-    moment(date).format("MMM D, YYYY") === moment(today).format("MMM D, YYYY")
-  const format = hasYear ? "MMM D, YYYY" : "MMM D"
-
   return (
     <NewsNavContainer>
       <MaxWidthContainer>
-        <NavText>{isToday ? "Today" : moment(date).format(format)}</NavText>
+        {date && <NewsDateHeader date={date} />}
         <Title>The News</Title>
       </MaxWidthContainer>
     </NewsNavContainer>
   )
 }
 
-const NavText = styled.div`
-  ${Fonts.unica("s25", "medium")};
-  ${pMedia.sm`
-    ${Fonts.unica("s16", "medium")}
-  `};
-`
-
-const Title = NavText.extend`
+const Title = NewsText.extend`
   position: absolute;
   left: 30px;
   ${pMedia.sm`
@@ -47,12 +34,19 @@ const MaxWidthContainer = styled.div`
   margin: auto;
   display: flex;
   justify-content: center;
+  align-items: center;
+  min-height: 26px;
+  ${pMedia.sm`
+    min-height: 16px;
+  `};
 `
 
 const NewsNavContainer = styled.div`
   position: fixed;
-  width: 100%;
-  box-shadow: 0 0 8px 0 rgba(0, 0, 0, 0.2);
+  top: 52px;
+  left: 0;
+  right: 0;
+  border-bottom: 1px solid ${colors.grayRegular};
   padding: 10px 0;
   background: white;
   z-index: 1;
